+++ src/PlatformSupport/XalanUTF16Transcoder.cpp
@@ -78,11 +78,11 @@ XalanUTF16Transcoder::~XalanUTF16Transcoder()
 XalanUTF16Transcoder::eCode
 XalanUTF16Transcoder::transcode(
 			const XalanDOMChar*		theSourceData,
-			unsigned int			theSourceCount,
+			size_t					theSourceCount,
 			XalanXMLByte*			theTarget,
-			unsigned int			theTargetSize,
-			unsigned int&			theSourceCharsTranscoded,
-			unsigned int&			theTargetBytesUsed)
+			size_t					theTargetSize,
+			size_t&					theSourceCharsTranscoded,
+			size_t&					theTargetBytesUsed)
 {
 	unsigned int	theSourceEaten = 0;
 	unsigned int	theTargetPosition = 0;
@@ -125,11 +125,11 @@ XalanUTF16Transcoder::transcode(
 XalanUTF16Transcoder::eCode
 XalanUTF16Transcoder::transcode(
 			const XalanXMLByte*		theSourceData,
-			unsigned int			theSourceCount,
+			size_t					theSourceCount,
 			XalanDOMChar*			theTarget,
-			unsigned int			theTargetSize,
-			unsigned int&			theSourceCharsTranscoded,
-			unsigned int&			theTargetBytesUsed,
+			size_t					theTargetSize,
+			size_t&					theSourceCharsTranscoded,
+			size_t&					theTargetBytesUsed,
 			unsigned char*			theCharSizes)
 {
 	unsigned int	theSourceEaten = 0;
